pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ms-ai-agent-framework"
version = "0.1.0"
description = "A unified framework to build and deploy agents using AutoGen, Semantic Kernel, and Azure AI Agent Service"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "click>=8.1",
    "structlog>=24.0",
    "opentelemetry-sdk>=1.20",
    "opentelemetry-api>=1.20",
    "watchfiles>=0.21",
]

[project.optional-dependencies]
autogen = [
    "autogen-agentchat>=0.4",
    "autogen-ext[openai]>=0.4",
]
semantic-kernel = [
    "semantic-kernel>=1.0",
]
azure = [
    "azure-ai-projects>=1.0.0b4",
    "azure-identity>=1.15",
    "azure-mgmt-containerinstance>=10.0",
    "azure-keyvault-secrets>=4.7",
]
docker = [
    "docker>=7.0",
]
ui = [
    "gradio>=4.0",
]
all = [
    "ms-ai-agent-framework[autogen,semantic-kernel,azure,docker,ui]",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.12",
]

[project.scripts]
agent = "agent_framework.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["agent_framework"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
